mapas funcionando

+++ app/src/main/java/com/svalero/musicrightsapp/contract/RegisterConcertContract.java
@@ -14,7 +14,7 @@ public interface RegisterConcertContract    {
     }
 
     interface Presenter {
-        void registerConcert(String showTitle, String city, String province, LocalDate date, String status, Boolean performed, Float ticketPrice, float latitude, float longitude, Musician musician);
+        void registerConcert(String showTitle, String city, String province, LocalDate date, String status, Boolean performed, Float ticketPrice, Double latitude, Double longitude, Musician musician);
         void modifyConcert(long id, String showTitle, String city, String province, LocalDate date, String status, Boolean performed, Float ticketPrice);
 }
